Split `Throttle` implementation into modules
